4.6*: Helper Functions and Unit Tests, step 4
Finishing this exercise can be done without the use of additional libraries.
However, this exercise is a great opportunity to learn how to use the Lodash library.

Define a function called mostBlogs that receives an array of blogs as a parameter.
The function returns the author who has the largest amount of blogs.
The return value also contains the number of blogs the top author has.

Signed-off-by: sharifmdathar <116189751+sharifmdathar@users.noreply.github.com>