4.9: Blog List Tests, step 2
Write a test that verifies that the unique identifier property of the blog posts is named id, by default the database names the property _id.

Make the required changes to the code so that it passes the test. The toJSON method discussed in part 3 is an appropriate place for defining the id parameter.

Signed-off-by: sharifmdathar <116189751+sharifmdathar@users.noreply.github.com>