4.10: Blog List Tests, step 3
Write a test that verifies that making an HTTP POST request to the /api/blogs URL successfully creates a new blog post. At the very least, verify that the total number of blogs in the system is increased by one. You can also verify that the content of the blog post is saved correctly to the database.

Once the test is finished, refactor the operation to use async/await instead of promises.

Signed-off-by: sharifmdathar <116189751+sharifmdathar@users.noreply.github.com>