4.12*: Blog List tests, step 5
Write tests related to creating new blogs via the /api/blogs endpoint, that verify that if the title or url properties are missing from the request data, the backend responds to the request with the status code 400 Bad Request.

Make the required changes to the code so that it passes the test.

Signed-off-by: sharifmdathar <116189751+sharifmdathar@users.noreply.github.com>