4.20*: Blog List Expansion, step 8 This example from part 4 shows taking the token from the header with the getTokenFrom helper function in controllers/blogs.js.
If you used the same solution, refactor taking the token to a middleware. The middleware should take the token from the Authorization header and assign it to the token field of the request object.

In other words, if you register this middleware in the app.js file before all routes

Routes can access the token with request.token:

Remember that a normal middleware function is a function with three parameters, that at the end calls the last parameter next to move the control to the next middleware:

Signed-off-by: sharifmdathar <116189751+sharifmdathar@users.noreply.github.com>