4.22*: Blog List Expansion, step 10
Both the new blog creation and blog deletion need to find out the identity of the user who is doing the operation. The middleware tokenExtractorthat we did in exercise 4.20 helps but still both the handlers of post and delete operations need to find out who the user holding a specific token is.

Now create a new middleware userExtractor, that finds out the user and sets it to the request object. When you register the middleware in app.js
the user will be set in the field request.user

Note that it is possible to register a middleware only for a specific set of routes. So instead of using userExtractor with all the routes,
we could register it to be only executed with path /api/blogs routes:

As can be seen, this happens by chaining multiple middlewares as the arguments of the function use. It would also be possible to register a middleware only for a specific operation:

Signed-off-by: sharifmdathar <116189751+sharifmdathar@users.noreply.github.com>