5.1: Blog List Frontend, step 1
Implement login functionality to the frontend. The token returned with a successful login is saved to the application's state user.

If a user is not logged in, only the login form is visible.

If the user is logged-in, the name of the user and a list of blogs is shown.

User details of the logged-in user do not have to be saved to the local storage yet.

Signed-off-by: sharifmdathar <116189751+sharifmdathar@users.noreply.github.com>