5.5-5.6: Complete Multiple Exercises
5.5 Blog List Frontend, step 5
Change the form for creating blog posts so that it is only displayed when appropriate. Use functionality similar to what was shown earlier in this part of the course material. If you wish to do so, you can use the Togglable component defined in part 5.

By default the form is not visible
It expands when button create new blog is clicked
The form hides again after a new blog is created.

5.6 Blog List Frontend, step 6
Separate the form for creating a new blog into its own component (if you have not already done so), and move all the states required for creating a new blog to this component.

The component must work like the NoteForm component from the material of this part.

Signed-off-by: sharifmdathar <116189751+sharifmdathar@users.noreply.github.com>