5.12: Blog List Frontend, step 12
Define PropTypes for one of the components of your application, and add ESlint to the project. Define the configuration according to your liking. Fix all of the linter errors.

Vite has installed ESlint to the project by default, so all that's left for you to do is define your desired configuration in the .eslintrc.cjs file.

Signed-off-by: sharifmdathar <116189751+sharifmdathar@users.noreply.github.com>